Extract schedule rideshare and assignment panels
